Stricter standards
Upgraded the makefile to include a lot more warnings and to be
stricter when dealing with them

When reading from the socket, it is now guaranteed that a specific
amount of bytes WILL be read (before it was possible that the call
would read less bytes than requested.)

+++ src/btea.c
@@ -6,9 +6,9 @@
 // https://en.wikipedia.org/wiki/XXTEA
 
 #define DELTA 0x9e3779b9
-#define MX (((z>>5^y<<2) + (y>>3^z<<4)) ^ ((sum^y) + (Identification_Key[(p&3)^e] ^ z)))
+#define MX (((z >> 5 ^ y << 2) + (y >> 3 ^ z << 4)) ^ (( sum ^ y) + ( Identification_Key[(p & 3) ^ e] ^ z)))
 
-void btea(uint32_t *v, int n)
+void btea(uint32_t *v, size_t n)
 {
 	uint32_t y, z, sum;
 	unsigned int p, rounds, e;
@@ -18,11 +18,11 @@ void btea(uint32_t *v, int n)
 	do {
 		sum += DELTA;
 		e = (sum >> 2) & 3;
-		for (p=0; p<n-1; p++) {
-			y = v[p+1]; 
+		for (p = 0; p < n - 1; p++) {
+			y = v[ p + 1]; 
 			z = v[p] += MX;
 		}
 		y = v[0];
-		z = v[n-1] += MX;
+		z = v[n - 1] += MX;
 	} while (--rounds);
 }
