+++ src/btea.h
@@ -1,3 +1,4 @@
 #include <stdint.h>
+#include <stdlib.h>
 
-void btea(uint32_t *v, int n);
+void btea(uint32_t *v, size_t n);
